experiment: diff font algorithm
Core premise was trying to equalize based on cap height and a lot of
downstream tweaks to layout algo, ex. using ParagraphLayout. It seemed
visibly better, more consistent. But this also seemed due to
force-setting a line height of 1.33, and most importantly, the default
body:label font ratio felt wrong with the same font size.