scaffold: try setting fill to match contrast
Currently, if color exceeds required contrast, fill is color.

This turns out to be disruptive to illustrations. Their visual hierarchy ends up out of whack if you can't guarantee the relative contrast of fill vs. text.

Additionally, in retrospect, it was a bad idea. It was a workaround when `color` support hadn't advanced enough to be able to easily guarantee contrast in all sorts of scenarios (for example, shadows & border)

Now that we understand colorHover is the best for glow, and UI components like MonetElevatedButton make using color and enforcing contrast easy, color is the correct choice when you want to have the direct theme color and still have contrast.

Thus, fill can go back to being a more understandable role, which in the short term resolves issues with building from tone, illustrations, and dynamic color.