Revert "fix: MonetTheme injects DefaultTextStyle"
This reverts commit b6ea79d600a7e903fb70dc1adec257ffe4101f2b.

This causes issues.

For instance, client used to use DefaultTextStyle.of(context).color in
an ElevatedButton's child that was a button itself to obtain the proper
hover/splash color.

Honestly, I don't know why that worked, but it does.

However, it stopped working after this change. And debugging was
difficult - they added a DefaultTextStyle immediately above the button,
but in between the public button interface and the interior private
button, a MonetTheme was added, injecting a DefaultTextStyle.