fix: hct x opacity
Somehow, only on web, we observed the loading background in Telosnex wasn't visible w/o wallpaper.

This sat for a month or two.

Eventually, it was traced back to the result from lerpKeepHue always had an opacity of 0.003.

It's unclear why this *only* failed on web. It must have something to do with alphaNeue / the deprecated alpha method behaving differently on web.